♻️  Rename and move the sidebar footer
