+++ frontend/lib/components/Sidebar/components/SidebarFooter/components/UserButton.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 
 import { useSupabase } from "@/lib/context/SupabaseProvider";
 
-import { useGravatar } from "../hooks/useGravatar";
+import { useGravatar } from "../../../../../hooks/useGravatar";
 import { sidebarLinkStyle } from "../styles/SidebarLinkStyle";
 
 export const UserButton = (): JSX.Element => {
