+++ frontend/lib/hooks/useGravatar.ts
